修复时间24小时与12小时混用的bug

+++ src/main/java/net/xiayule/netbar/utils/Utils.java
@@ -8,12 +8,13 @@ import java.util.Calendar;
 
 public class Utils {
 
-	//根据下机时间和上机时间计算费用
+	/**
+	 * 根据下机时间和上机时间计算费用, 不够一小时，算一小时
+	 */
 	public static double calc(Calendar begintime, Calendar endtime) {
-		return (double) (Math.ceil((endtime.getTimeInMillis() - begintime
-				.getTimeInMillis())/1000.0/60/60)*2);
+		return Math.ceil((endtime.getTimeInMillis() - begintime
+				.getTimeInMillis()) / 1000.0 / 60 / 60)*2;
 	}
-
 	
 	//显示警告对话框
 	public  static void showDialog(String s){
